Added webpack to bundle helpers script Fixes #45
- Import helpers.js in project to use defineCustomElements('component-name')
- See the demo in storybook setup
