+++ package.json
@@ -22,7 +22,8 @@
     "lint-fix": "npm run lint -- --fix",
     "prettify": "prettier {*,src/**/*,.storybook/**/*}.{js,jsx,ts,tsx,json,css,scss,md,html} --write",
     "test": "stencil test --spec --e2e",
-    "test.watch": "stencil test --spec --e2e --watchAll"
+    "test.watch": "stencil test --spec --e2e --watchAll",
+    "webpack": "webpack"
   },
   "husky": {
     "hooks": {
@@ -42,7 +43,8 @@
   "devDependencies": {
     "@babel/core": "^7.2.2",
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/preset-env": "^7.2.3",
+    "@babel/plugin-transform-modules-commonjs": "^7.2.0",
+    "@babel/preset-env": "^7.3.4",
     "@babel/preset-typescript": "^7.1.0",
     "@fortawesome/fontawesome-free": "^5.6.3",
     "@stencil/core": "^0.18.0",
@@ -53,9 +55,11 @@
     "@storybook/html": "^4.1.11",
     "@types/jest": "23.3.13",
     "@types/puppeteer": "1.12.1",
-    "babel-core": "7.0.0-bridge.0",
-    "babel-loader": "^8.0.4",
+    "babel-core": "^7.0.0-bridge.0",
+    "babel-loader": "^8.0.5",
+    "babel-preset-env": "^1.7.0",
     "bootstrap": "^4.3.1",
+    "clean-webpack-plugin": "^1.0.1",
     "css-loader": "^2.1.0",
     "html-loader": "^0.5.5",
     "http-proxy-middleware": "^0.19.1",
@@ -75,7 +79,8 @@
     "tslint": "^5.12.1",
     "tslint-config-prettier": "^1.18.0",
     "url-loader": "^1.1.2",
-    "webpack": "4.28.4"
+    "webpack": "4.28.4",
+    "webpack-cli": "^3.2.3"
   },
   "license": "MIT",
   "dependencies": {}
